cleanup header files and add util folder
