+++ src/widgets/desktops.h
@@ -1,3 +1,7 @@
+#include <string.h>
+#include <stdbool.h>
+#include <xcb/xcb_icccm.h>
+
 #define PROPERTY_MAX_LEN 256
 #define DESKTOP_MAX_LEN 10
 
@@ -7,7 +11,3 @@ typedef struct desktop_t {
 	bool is_valid;
 	int clients_len;
 } desktop_t;
-
-static int ewmh_get_active_window_name(xcb_ewmh_connection_t *ewmh, int screen_nbr, char *window_name);
-static int ewmh_get_desktop_list(xcb_ewmh_connection_t *ewmh, int screen_nbr, desktop_t *desktops);
-static int widget_desktops_send_update ();
