+++ src/widgets/now_playing_mpd.c
@@ -1,5 +1,3 @@
-#include "../wkline.h"
-#include "../config.h"
 #include "widgets.h"
 #include "now_playing_mpd.h"
