+++ src/widgets/volume.c
@@ -1,5 +1,3 @@
-#include "../wkline.h"
-#include "../config.h"
 #include "widgets.h"
 #include "volume.h"
