+++ src/widgets/volume.h
@@ -1,3 +1,4 @@
+#include <webkit/webkit.h>
 #include <alsa/asoundlib.h>
 
-static int widget_volume_send_update (snd_mixer_elem_t* elem);
+#include "src/config.h"
