+++ src/wkline.c
@@ -1,8 +1,8 @@
 #include "config.h"
 #include "wkline.h"
+#include "util/curl.h"
+#include "util/log.h"
 #include "widgets/widgets.h"
-#include "widgets/desktops.h"
-#include "lib/curl.h"
 
 thread_data_t thread_data;
 WebKitWebView *web_view;
@@ -44,24 +44,6 @@ wk_notify_load_status_cb (WebKitWebView *web_view, GParamSpec *pspec, GtkWidget 
 			}
 		}
 	}
-}
-
-void
-wklog (char const *format, ...) {
-#ifdef DEBUG
-	va_list args;
-	time_t rawtime;
-	struct tm *date;
-
-	time(&rawtime);
-	date = localtime(&rawtime);
-
-	fprintf(stderr, "[%02d:%02d:%02d] wkline: ", date->tm_hour, date->tm_min, date->tm_sec);
-	va_start(args, format);
-	vfprintf(stderr, format, args);
-	va_end(args);
-	fprintf(stderr, "\n");
-#endif
 }
 
 int
