+++ util/log.h
@@ -1,0 +1,4 @@
+#include <stdio.h>
+#include <time.h>
+
+void wklog (char const *format, ...);
